+++ packages/mdx/src/core.ts
@@ -294,7 +294,7 @@ function postprocessPlugin(): Plugin {
   /**
    * extracted references (e.g. hrefs, paths), useful for analyzing relationships between pages.
    */
-  extractedReferences?: import('fumadocs-mdx').ExtractedReference[];
+  extractedReferences: import("fumadocs-mdx").ExtractedReference[];
 }`;
 
   return {